quiet warnings with casts

+++ src/xftname.c
@@ -36,7 +36,7 @@ XftNameUnparse (FcPattern *pat, char *dest, int len)
     name = FcNameUnparse (pat);
     if (!name)
 	return FcFalse;
-    if (strlen ((char *) name) + 1 > len)
+    if (strlen ((char *) name) + 1 > (size_t) len)
     {
 	FcPattern *new = FcPatternDuplicate (pat);
 	free (name);
@@ -46,7 +46,7 @@ XftNameUnparse (FcPattern *pat, char *dest, int len)
 	FcPatternDestroy (new);
 	if (!name)
 	    return FcFalse;
-	if (strlen ((char *) name) + 1 > len)
+	if (strlen ((char *) name) + 1 > (size_t) len)
 	{
 	    strncpy (dest, ((char *) name), (size_t) len - 1);
 	    dest[len - 1] = '\0';
